pyproject.toml
[tool.poetry]
name = "motoman"
version = "0.1.0"
description = "A command line tool to control a set of stepper motors"
license = "MIT"
authors = ["Satyam Tiwary <satyamtiwary@vvbiotech.com>"]
readme = 'README.md'
repository = "https://github.com/TechnocultureResearch/MotorsManager-CLI"
homepage = "www.technocultureresearch.org"
keywords = ['instrumentation', 'control', 'motors', 'serial']
classifiers = [
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: System :: Hardware :: Universal Serial Bus (USB)",
    "Environment :: MacOS X"
]

[tool.poetry.dependencies]
python = "^3.7"
toml = "^0.9"
prompt-toolkit = "^3.0.17"
Pygments = "^2.8.1"
python-nubia = "^0.2b5"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
pytest = "^6.0"
pytest-cov = "^2.4"

[tool.poetry.scripts]
motoman = 'motoman:main' # main function of motoman.py -> Entry point

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
